+++ sdk/src/connections/offline/offline_depth_sensor.h
@@ -87,7 +87,7 @@ class OfflineDepthSensor : public aditof::DepthSensorInterface {
                                                int &imagerStatus) override;
     virtual aditof::Status
     initTargetDepthCompute(uint8_t *iniFile, uint16_t iniFileLength,
-                           uint8_t *calData, uint16_t calDataLength) override;
+                           uint8_t *calData, uint32_t calDataLength) override;
     virtual aditof::Status
     getDepthComputeParams(std::map<std::string, std::string> &params) override;
